Remove Angular
Angular was too restrictive and not a good conceptual fit for what I
want to do, so I removed it and replaced it with direct DOM manipulation,
which isn't too bad in Dart if you use the cascade notation as much as
possible.

This commit has (as far as I can tell) the same functionality as the
previous commit, so this is a pure refactoring commit.