Add "Warehouse" feature
This replaces the list of assemblies in the previous version of
Encoding.Tools. The warehouse lists prebuilt assemblies of gadgets and
is based on a JSON-friendly object format to facilitate a future feature
where gadgets can be saved and loaded again later.

The prebuilt assemblies no longer include one entry for each transform,
since that wasn't very DRY. Now the user can double click on a transform
to construct an assembly automatically.